backend/requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
python-dotenv
rich
asgi-correlation-id
python-json-logger
python-jose
python-multipart
passlib[bcrypt]
aiofiles
b2sdk
PyMuPDF  
redis
transformers>=4.0.0
torch>=2.0.0  # CPU-only PyTorch
transformers
numpy>=1.20
faiss-cpu>=1.7
sentence-transformers>=2.2